Weave Resonance Fix
The spellhook was looking at a strange feat id when determining whether to give feedback on the state on the Weave upon casting a spell. As far as I can tell, the feat was blank. Seems like the intention here was to check for the Weave Resonance feat (1499), since this is literally what that feat does.

I've also updated the feedback strings to make them a little less clunky (and a little less like some sort of OOC error message).

The spell failure message due to dead magic indicated that the problem was you not having enough spellcraft, which is confusing, because there's no amount of spellcraft skill that will tell you anything different here. So changed it to a line about how "the magic unravels".