Fix Check for Survival Proficiency On Rest
The WeatherTest function returns TRUE if you can't rest due to ambient weather conditions.

It was checking for Feat 1439, and returning FALSE if the feat was present. But this is the feat id for the Tracking proficiency.

Corrected the typo to point towards the correct feat id for the Survival proficiency.