freshening up weave chat command
Bringing the chat command afforded by the Weave Resonance feat into line with the behaviour of Weave Resonance feedback in the spellhook.

Added an extra two minor cases at <1 - 9> and <10-32> so as to make the feedback less misleading.

Previous behaviour: the Weave is clearly broken, "The weave here is fairly healthy"

New behaviour: if there's any chance of spell failure due to Weave damage, you get a message about the Weave being disrupted.