fix check weave resonance feat on casting
The spellhook referenced a feat that appeared not to exist, and if found would give some information about the state of the local Weave. Since that's literally the function of the Weave Resonance feat, corrected the feat id to refer to that. This should fix some more of the feat's functionality.

Punched up the bland descriptions a little.

The spell failure notification said you need more Spellcraft, but the Spellcraft skill does not help you here. Some players were even interpreting this as a debug message. So made it a little more IC-sounding while preserving the ambiguity.

Added some comments.