tracking animation fix
It's pretty obvious that the intent here was to make the animation last for six seconds, and then as the PC is standing back up they receive the results. But they passed the 6.0 into the speed parameter and the duration parameter defaulted to 1.0, so a PC doing a tracking bends down for a sixth of a second and then springs back up, then gets the results six seconds later for some reason. It should work properly now, though we might also want to consider rooting the PC in place like with bandages and chests.