Free Cantrips
An event handler has been created (subscribed on mod load) for when spell slots are consumed, which will refund level 0 cantrips cast by wizards and sorcerers. It handles multiclassing and metamagic gracefully: only cantrip slots will ever be refunded, and only the cantrips wizards prepared at the start of the adventuring day are unlimited.

An ImpactScript has been created for the spell Disrupt Undead. It works as the spell description describes.

Cantrips no longer contribute to weave corruption or repair. Spell level is passed through the functions in the x2 spellhook into loi_weave. Weave state changes only happen if that is > 0

The unuse script tb_spellhook has been deleted.

A SpellFizzle function has been created in loi_weave for when dead magic causes a spell to fail, which looks up the casting animation type and plays the appropriate "fizzle" VFX, as if the spell had been interrupted. It is called in the x2 spellhook alongside the dead magic function.