The Ankhegs Must Die
The script that causes Paladins to lose piety and lawful alignment if they kill hostile Ankhegs or animals has been modified to also exclude animals, beasts, magical beasts, oozes, aberrations, and elementals, all of which could be incapable of moral reasoning.

Added a feedback message that tells the player why this is happening.

Set the transgression flag in the case of a paladin falling for killing a Good-aligned NPC, because that behaviour is inconsistent and buggy otherwise.

Optimised slightly, so it only needs to look up the slain creature's racial type once.