Fix Rumors
Using DelayCommand() with a 24 hour irl timer is horrifying: if the PC logs off or dies or changes state who knows what happens.

Rewritten the check so it instead uses the calendar day (1 - 28). If you already asked on the exact same day, no dice. There's an edge case where you ask on the same day *next* month, but if so that's just bad luck - try again the next day.

You get the mythic exp if you succeed the check, but can no longer farm the check.

The persuasion check is now visible to the player.

We declare all our variables ahead of time, to make it clearer. We also declare our DC, to make it simpler later to retrieve the DC dynamically from the NPC.